+++ server/modules/doc_handler.js
@@ -8,8 +8,8 @@
 import winston from 'winston';
 import PostgresStore from './postgres_store';
 import KeyGenerator from './key_generator';
-import Config from '../../config';
-import { privacyOptions } from '../config/constants';
+import Config from '../config/config';
+import { privacyOptions } from '../../shared/config/constants';
 import { postgresTimestamp } from './_helpers';
 
 class DocHandler {